# yuanzi-backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "小园子母婴记录应用后端 API (基于 PRD-V1.1)",
        "title": "Yuanzi API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/stats/daily": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取指定日期的喂养/睡眠/排泄统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "日统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "宝宝ID",
                        "name": "baby_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期(YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.DailyStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/stats/weekly": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取近7天喂养/睡眠/排泄统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "周统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "宝宝ID",
                        "name": "baby_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.WeeklyStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/sync/stream": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "建立服务器发送事件连接，接收家庭数据变更",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "同步"
                ],
                "summary": "SSE 实时同步",
                "responses": {
                    "200": {
                        "description": "SSE 连接建立"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.DailyStatsResponse": {
            "type": "object",
            "properties": {
                "diaper": {
                    "$ref": "#/definitions/handler.DiaperStats"
                },
                "feeding": {
                    "$ref": "#/definitions/handler.FeedingStats"
                },
                "sleep": {
                    "$ref": "#/definitions/handler.SleepStats"
                }
            }
        },
        "handler.DiaperStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "handler.FeedingStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "integer"
                }
            }
        },
...
        "handler.SleepStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "total_hours": {
                    "type": "number"
                }
            }
        },
...
        "handler.WeeklyStatsResponse": {
            "type": "object",
            "properties": {
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "diaper": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "feeding": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sleep": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}